Trait mempool_notifications::MempoolNotificationSender

pub trait MempoolNotificationSender: Send + Clone + Sync + 'static {
    fn notify_new_commit<'life0, 'async_trait>(
        &'life0 self, 
        committed_transactions: Vec<Transaction>, 
        block_timestamp_usecs: u64, 
        notification_timeout_ms: u64
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

The interface between state sync and mempool, allowing state sync to notify mempool of events (e.g., newly committed transactions).

Required methods

fn notify_new_commit<'life0, 'async_trait>(
&'life0 self,
committed_transactions: Vec<Transaction>,
block_timestamp_usecs: u64,
notification_timeout_ms: u64
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Notify mempool of the newly committed transactions at the specified block timestamp.